fix errors on windows (hex castings and warning)

+++ src/core/control/settings/Settings.h
@@ -53,7 +53,7 @@ public:
     virtual ~SAttribute();
 
     std::string sValue;
-    int iValue{};
+    long iValue{};
     double dValue{};
 
     AttributeType type;
@@ -114,7 +114,7 @@ private:
 
     static xmlNodePtr savePropertyDouble(const char* key, double value, xmlNodePtr parent);
     static xmlNodePtr saveProperty(const char* key, int value, xmlNodePtr parent);
-    static xmlNodePtr savePropertyUnsigned(const char* key, unsigned int value, xmlNodePtr parent);
+    static xmlNodePtr savePropertyUnsigned(const char* key, size_t value, xmlNodePtr parent);
     static xmlNodePtr saveProperty(const char* key, const char* value, xmlNodePtr parent);
 
     void saveData(xmlNodePtr root, const std::string& name, SElement& elem);
